Rename core.qcstate to core.dataStates for more general use
