+++ core/resources/schemas/dbscripts/postgresql/core-21.005-21.006.sql
@@ -1,0 +1,3 @@
+ALTER TABLE core.qcstate RENAME TO dataStates;
+
+ALTER TABLE core.dataStates ADD COLUMN stateType VARCHAR(20);
