+++ core/resources/schemas/dbscripts/sqlserver/core-21.005-21.006.sql
@@ -1,0 +1,4 @@
+EXEC sp_rename 'core.qcstate', 'dataStates'
+GO
+
+ALTER TABLE core.dataStates ADD stateType NVARCHAR(20);
